ewm-server 0.6

+++ ewm-service/src/main/java/ru/practicum/exploreWithMe/privateApi/service/PrivateServiceImpl.java
@@ -295,7 +295,7 @@ public class PrivateServiceImpl implements PrivateService {
     }
 
     private void ownerEventVerification(Long eId, Long uId) {
-        if (eId != uId) {
+        if (eId.longValue() != uId.longValue()) {
             throw new BadRequest("нельзя изменить чужие ивенты");
         }
     }
